+++ test/IntegrationSpec.hs
@@ -1,25 +1,18 @@
 module IntegrationSpec where
-import qualified Data.Text                     as T
-import           Database.MongoDB        hiding ( Key )
+
 import           Test.Hspec
 import           Test.Hspec.Wai
 import           Network.Wai                    ( Middleware )
-import           Model
-import           ModelUtils                     ( findObjects
-                                                , insertObject
-                                                , findById
-                                                , run
-                                                )
-import           BSONExtention
+
 import           Utils                          ( flushDB )
-import           Data.ByteString.Lazy.Internal  ( ByteString )
+
 import           Init                           ( app )
 import           Web.Spock                      ( spock
                                                 , spockAsApp
                                                 )
 import           Web.Spock.Config
-
-import Integration.AuthHandle (authHandleTest)
+----------------------------------------------------------------
+import Integration.AuthTest (authTest)
 
 beforeAll = do
   flushDB
@@ -37,4 +30,4 @@ main = hspec spec
 
 spec :: Spec
 spec = with (spockAsApp testApp) $ do
-  authHandleTest
+  authTest
